+++ primint/src/num/nonzero.rs
@@ -86,6 +86,12 @@ impl<T: PrimitiveInt> NonZero<T> {
     ///
     /// Always equals the number of bits of the underlying type.
     pub const BITS: u32 = crate::bits::<T>();
+
+    /// The constant number `1`.
+    pub const ONE: Self = {
+        // SAFETY: One is never zero
+        unsafe { Self::new_unchecked(T::ONE) }
+    };
 }
 impl<T: PrimitiveInt> Display for NonZero<T> {
     fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {