+++ CHANGELOG.md
@@ -15,6 +15,7 @@ A title is required for publishing a github release, so all versions should have
 - Add infallible `From` conversions for `NonMax`, `NonZero` (xutqswvn)
   - No failable conversions yet, as that would require defining a `TryFromIntError` type.
 - Add `BITS` associated constant to `NonMax`, `NonZero` (rslwssvw)
+- Add `ONE`, `ZERO` constants to `NonMax` and `ONE` constant to `NonZero` (mkyykwvx)
 
 ## 0.1.4 - 2026-08-25
 Make `NonZero::new`, `NonMax::{new, get}` a `const fn`.